# api/gateway.openapi.yaml
...
openapi: 3.1.0
info:
  title: Multiverse gateway API
  version: 1.0.0
  description: |
    Player entry point of the platform: links and consent, characters, actions,
    groups, deliveries of results. Clients are the Telegram bot and the CI
    harness; both use internal/gateway/client.

    Every answer carries the header X-Request-Id (components.headers.XRequestId;
    set by the middleware of component §5.1, T-303), so operations do not repeat it.

    Every error answer has the body {"error": {"code", "message", "details"}}.
    A client matches on code; message is plain Russian text for the player.
servers:
  - url: http://127.0.0.1:8088
security:
  - ClientId: []
...
paths:
  /v1/links/resolve:
    post:
      operationId: resolveLink
      tags: [links]
      summary: Resolve an external account into a player
      description: |
        Creates a pending_consent link with a new link_id on the first call and
        updates last_seen_at on every call. link_id is not returned (SEC-03).
        notice_due is true for a new or pending_consent link and when the
        previous visit was 30 days ago or more. character_status is none while
        the link has no character. external_platform other than telegram and an
        empty external_id answer 400 invalid_request.
      x-nolog: true
      parameters:
        - $ref: '#/components/parameters/ActorKind'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveRequest' }
      responses:
        '200':
          description: The state of the link and of the character.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResolveResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/links/consent:
    post:
      operationId: consentLink
      tags: [links]
      summary: Record the notice, the consent and the age confirmation
      description: |
        Any of the three flags not true answers 400 consent_incomplete: an
        existing link stays pending_consent, and an account without a link
        gets none (nothing stores its external ID). A complete consent sets
        notice_shown_at to shown_at and consent_at, age_confirmed_at to the time
        of the gateway; a repeat keeps the first three times (they are the
        evidence of the consent) and only updates last_seen_at. A complete
        consent for an account that was never resolved creates its link.
        shown_at missing answers 400 invalid_request.
      x-nolog: true
      parameters:
        - $ref: '#/components/parameters/ActorKind'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsentRequest' }
      responses:
        '200':
          description: The link is consented.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsentResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/links:
    delete:
      operationId: forgetLink
      tags: [links]
      summary: Delete the link (/forget)
      description: |
        Cascade of C-04 v1.3 in this order: the outbox of the player is dropped;
        a living character gets status=abandoned (and a new leader or none for
        its group), group.left {cause: forget}; the active session ends with
        end_reason=forget; the link is deleted physically. A step that fails
        stops the cascade before the deletion, so a repeat runs it again.

        What is deleted (SEC-26, threat-model §4.8): the row of the link — the
        external ID, player_id, world_id, the times of the notice, the consent
        and the age confirmation, last_seen_at, the internal link_id — and the
        stored answers of POST /v1/characters under that link. The deletion is
        physical before the answer: secure_delete zeroes the cells, the free
        pages are vacuumed and the WAL is checkpointed and truncated, so the
        external ID is left neither in the file nor in its WAL, and nothing
        finds the external account by player_id any more.

        What stays: the character under its pseudonymous player_id (abandoned,
        or dead) in the world state and snapshots; the game history by player_id
        in the event journal until its retention (30 days, LLM records 90,
        analytics 180) and in the memory indexes built from it; sessions and
        turns of gateway.db by player_id; the encrypted backup of the links file
        for at most 30 days. A new /start after /forget is a new player with a
        new player_id.

        Answers: 200 {deleted: true, player_id_detached} when a link was deleted,
        player_id_detached null when it had no character (the client says there
        was nothing to delete); 200 {deleted: false, player_id_detached: null}
        without a link, which is also the answer to a repeat of a /forget that
        already succeeded.

        503 forget_incomplete, unlike bus_unavailable, does not mean that nothing
        was done: the link is deleted, but the wipe of its bytes from the file of
        links is not confirmed. Causes: a reader outside the gateway, such as a
        backup, holding the file past the busy timeout, or the deadline of the
        request. The data must not be reported as gone. While that wipe stays
        pending, every /forget answers 503 forget_incomplete, the repeats of the
        same account and the /forget of other accounts, with or without a link;
        the call that finishes the wipe answers as usual. The answer carries
        Retry-After; the client repeats, and a 200 {deleted: false} after a 503
        is the end of the same /forget, not "nothing to delete". The gateway
        also finishes the wipe itself: its sweeper retries every minute, and
        every start of the gateway compacts the file of links.
      x-nolog: true
      parameters:
        - $ref: '#/components/parameters/ActorKind'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgetRequest' }
      responses:
        '200':
          description: What was removed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ForgetResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    ClientId:
      type: apiKey
      in: header
      name: X-Client-Id
      description: Client instance id; must be listed in MV_GATEWAY_CLIENT_IDS, otherwise 403 client_unknown (C-08 v1.3).
  parameters:
    ActorKind:
      name: X-Actor-Kind
      in: header
      required: false
      description: Kind of actor; ci and sim only for clients listed in MV_GATEWAY_ACTOR_KIND_CLIENTS, otherwise 403 actor_kind_forbidden.
      schema:
        type: string
        enum: [human, ci, sim]
        default: human
...
  responses:
    BadRequest:
      description: 'Request does not match the contract. Codes: invalid_request, unknown_action, text_invalid, name_required, name_invalid, consent_incomplete.'
      x-error-codes: [invalid_request, unknown_action, text_invalid, name_required, name_invalid, consent_incomplete]
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: invalid_request, message: Некорректный запрос. } }
    Forbidden:
      description: 'Client or actor not allowed. Codes: consent_required, actor_kind_forbidden, client_unknown, client_mismatch.'
      x-error-codes: [consent_required, actor_kind_forbidden, client_unknown, client_mismatch]
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: client_unknown, message: Клиент не допущен. } }
...
    PayloadTooLarge:
      description: 'Body over 64 KiB (SEC-11). Codes: payload_too_large.'
      x-error-codes: [payload_too_large]
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: payload_too_large, message: Слишком большой запрос. } }
...
    InternalError:
      description: 'Defect of the gateway (NFR-012). Codes: internal.'
      x-error-codes: [internal]
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: internal, message: Внутренняя ошибка. } }
...
    Unavailable:
      description: |
        A dependency is down; the action is not accepted and action_key is not stored, the client repeats with the same key.
        Codes: bus_unavailable, state_unavailable, forget_incomplete (DELETE /v1/links: the link is deleted, the wipe
        from the file of links is not confirmed; the data is not reported as gone, the client repeats /forget after
        Retry-After, see forgetLink).
      x-error-codes: [bus_unavailable, state_unavailable, forget_incomplete]
      headers:
        Retry-After:
          description: Sent with forget_incomplete only; seconds before the client repeats /forget.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: bus_unavailable, message: 'Сервис временно недоступен, повторите.' } }
  schemas:
    Error:
      type: object
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
      required: [error]
    ErrorBody:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
      required: [code, message]
    ResolveRequest:
      type: object
      properties:
        external_platform: { type: string, enum: [telegram] }
        external_id: { type: string, minLength: 1 }
      required: [external_platform, external_id]
    ResolveResponse:
      type: object
      properties:
        link_status: { type: string, enum: [none, pending_consent, consented] }
        player_id: { type: [string, 'null'] }
        world_id: { type: [string, 'null'] }
        character_status: { type: string, enum: [none, creating, alive, dead] }
        notice_due: { type: boolean }
      required: [link_status, player_id, world_id, character_status, notice_due]
    ConsentRequest:
      type: object
      properties:
        external_platform: { type: string, enum: [telegram] }
        external_id: { type: string, minLength: 1 }
        notice_shown: { type: boolean }
        consent: { type: boolean }
        age_confirmed: { type: boolean }
        shown_at: { type: string, format: date-time }
      required: [external_platform, external_id, notice_shown, consent, age_confirmed, shown_at]
    ConsentResponse:
      type: object
      properties:
        link_status: { type: string, enum: [consented] }
        consent_at: { type: string, format: date-time }
        age_confirmed_at: { type: string, format: date-time }
        notice_shown_at: { type: string, format: date-time }
      required: [link_status, consent_at, age_confirmed_at, notice_shown_at]
    ForgetRequest:
      type: object
      properties:
        external_platform: { type: string, enum: [telegram] }
        external_id: { type: string, minLength: 1 }
      required: [external_platform, external_id]
    ForgetResponse:
      type: object
      properties:
        deleted: { type: boolean }
        player_id_detached: { type: [string, 'null'] }
      required: [deleted, player_id_detached]
...